2D AprilTag autoalign command - NOT pose estimate based - Calculates the offset for the given april tag, then drives the robot to the desired offset - PID may need to be tuned for IRL robot; currently tuned for sim - TODO: Posed based auto align; combined autoalign (pose based to get close; ATag based to get exact)

+++ src/main/java/frc/robot/subsystems/vision/VisionIO.java
@@ -29,7 +29,7 @@ public interface VisionIO {
     }
 
     /** Represents the angle to a simple target, not used for pose estimation. */
-    record TargetObservation(long id, Rotation2d tx, Rotation2d ty) {}
+    record TargetObservation(int id, Rotation2d tx, Rotation2d ty) {}
 
     /** Represents a robot pose sample used for pose estimation. */
     record PoseObservation(
