+++ src/main/java/frc/robot/subsystems/vision/VisionIOLimelight.java
@@ -86,7 +86,7 @@ public class VisionIOLimelight implements VisionIO {
         // Update target observation
         inputs.latestTargetObservation =
             new TargetObservation(
-                tidSubscriber.get(),
+                (int) tidSubscriber.get(),
                 Rotation2d.fromDegrees(txSubscriber.get()),
                 Rotation2d.fromDegrees(tySubscriber.get())
             );
